requirement.txt
groq
sentence-transformers
faiss-cpu
rank-bm25
pypdf
pandas
python-dotenv
numpy